pkp/pkp-lib#8866 Use jobs to fix the old region codes after the upgrade
